OzWeathers/src/requirements.txt
jupyter==1.0.0
pandas==1.4.3
requests==2.28.1
SQLAlchemy==1.4.39
pyarrow==8.0.0
pg8000==1.29.1
pytest==7.1.2
pylint==2.14.4
pyyaml==6.0
Jinja2==3.1.2
schedule==1.1.0
boto3==1.24.91
botocore==1.27.80
graphlib==0.9.5
io==0.0.1
s3transfer==0.6.0
art==5.7
numpy==1.23.4